fixed material build problem, moved ReactiveFormModule to public/private module wrappers
