--- pouch-dbs/config.json
@@ -1,5 +1,0 @@
-{
-  "couchdb": {
-    "uuid": "e7a32db4-2f1b-42a1-9ef0-733c7be29142"
-  }
-}
