+++ src/app/modules/public/public.module.ts
@@ -1,43 +1,48 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes, CanActivate } from '@angular/router';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { LoginComponent } from './components/login/login.component';
 import { RegisterComponent } from './components/register/register.component';
 import { PublicComponent } from './public.component';
 import { AnonymousGuard } from '../../guards/anonymous.guard';
-import { MaterialModule } from './../material.module';
+import { MaterialModule } from '../material.module';
+
 
 const publicRoutes: Routes = [
   {
-    path: '',
+    path:      '',
     component: PublicComponent,
-  //   canActivate: [AnonymousGuard],
-    children: [
-        {
-          path: 'login',
-          component: LoginComponent
-        },
-        {
-          path: 'register',
-          component: RegisterComponent
-        },
-        {
-            path: '**',
-            redirectTo: 'dashboard',
-            pathMatch: 'full'
-        }
+    //   canActivate: [AnonymousGuard],
+    children:  [
+      {
+        path:      'login',
+        component: LoginComponent
+      },
+      {
+        path:      'register',
+        component: RegisterComponent
+      },
+      {
+        path:       '**',
+        redirectTo: 'dashboard',
+        pathMatch:  'full'
+      }
     ]
   }
 ];
 
 @NgModule({
-    declarations: [LoginComponent, RegisterComponent, PublicComponent],
-    imports: [
-        CommonModule,
-        MaterialModule,
-        RouterModule.forChild(publicRoutes)],
-    exports: [LoginComponent, RegisterComponent, PublicComponent],
-    providers: [],
+  declarations: [LoginComponent, RegisterComponent, PublicComponent],
+  imports:      [
+    CommonModule,
+    MaterialModule,
+    FormsModule,
+    ReactiveFormsModule,
+    RouterModule.forChild(publicRoutes)],
+  exports:      [LoginComponent, RegisterComponent, PublicComponent],
+  providers:    [],
 })
-export class PublicModule {}
+export class PublicModule {
+}
